+++ Runtime/UI/BrowserViews/ExplorerView.cs
@@ -60,6 +60,9 @@ namespace ModIO.UI
         /// <summary>Number of seconds per page transition.</summary>
         public float pageTransitionTimeSeconds = 0.4f;
 
+        /// <summary>The priority to focus the selectables.</summary>
+        public List<Selectable> onFocusPriority = new List<Selectable>();
+
         [Header("Events")]
         /// <summary>Event for notifying listeners of a change to displayed mods.</summary>
         public ModPageChanged onModPageChanged = null;
@@ -194,6 +197,9 @@ namespace ModIO.UI
 
         /// <summary>Is the view a root view or window view?</summary>
         bool IBrowserView.isRootView { get { return true; } }
+
+        /// <summary>The priority to focus the selectables.</summary>
+        List<Selectable> IBrowserView.onFocusPriority { get { return this.onFocusPriority; } }
 
         // ---------[ INITIALIZATION ]---------
         /// <summary>Initializes private members.</summary>
@@ -352,6 +358,7 @@ namespace ModIO.UI
                 return;
             }
 
+            // collect vars
             int pageSize = this.m_modPageContainer.itemLimit;
             if(pageSize < 0)
             {
@@ -360,13 +367,21 @@ namespace ModIO.UI
 
             int targetPageIndex = this.m_modPage.CalculatePageIndex() + pageDifferential;
             int targetPageProfileOffset = targetPageIndex * pageSize;
-
-            Debug.Assert(targetPageIndex >= 0);
-            Debug.Assert(targetPageIndex < this.m_modPage.CalculatePageCount());
-
+            int modPageCount = this.m_modPage.CalculatePageCount();
             int pageItemCount = (int)Mathf.Min(pageSize,
                                                this.m_modPage.resultTotal - targetPageProfileOffset);
 
+            // Bounds check
+            if(targetPageIndex >= modPageCount) { targetPageIndex = modPageCount-1; }
+            if(targetPageIndex < 0){ targetPageIndex = 0; }
+
+            // early out
+            if(targetPageIndex == this.m_modPage.CalculatePageIndex())
+            {
+                return;
+            }
+
+            // do transition
             RequestPage<ModProfile> transitionPlaceholder = new RequestPage<ModProfile>()
             {
                 size = pageSize,