+++ Runtime/UI/BrowserViews/InspectorView.cs
@@ -15,6 +15,9 @@ namespace ModIO.UI
         /// <summary>Id of the currently displayed mod.</summary>
         private int m_modId = ModProfile.NULL_ID;
 
+        /// <summary>The priority to focus the selectables.</summary>
+        public List<Selectable> onFocusPriority = new List<Selectable>();
+
         // --- ACCESSORS ---
         /// <summary>Id of the currently displayed mod.</summary>
         public int modId
@@ -80,6 +83,9 @@ namespace ModIO.UI
 
         /// <summary>Is the view a root view or window view?</summary>
         bool IBrowserView.isRootView { get { return false; } }
+
+        /// <summary>The priority to focus the selectables.</summary>
+        List<Selectable> IBrowserView.onFocusPriority { get { return this.onFocusPriority; } }
 
         // ---------[ UI Functionality ]---------
         /// <summary>Closes this view.</summary>